Migrate exceptions to separate file
